Extended to enable local model import as well
